add delete function for unvoted/uncommented posts. refactor editor to be a separate component. add edit redux action.
