+++ src/components/Editor/index.css
@@ -1,0 +1,9 @@
+.editor-tags {
+  margin-top: 10px;
+  margin-bottom: 10px;
+}
+
+.editor-rte {
+  margin-top: 10px;
+  margin-bottom: 10px;
+}
