--- src/containers/NewContribution/index.css
@@ -1,9 +1,0 @@
-.editor-tags {
-  margin-top: 10px;
-  margin-bottom: 10px;
-}
-
-.editor-rte {
-  margin-top: 10px;
-  margin-bottom: 10px;
-}
